back-end/requirements.txt
Django
django-ninja
django-ninja-jwt[crypto]
django-cors-headers
rav
